+++ demo-app/src/async-state-var/async-component-2.js
@@ -12,8 +12,8 @@ export class AsyncComponent2 extends LitStateElement {
 
             <h2>&lt;component-2&gt;</h2>
 
-            <h3>Status: ${this.dataStatus}</h3>
-            <h3>Value: ${demoState.data.getValue()}</h3>
+            <h3 id="status">Status: ${this.dataStatus}</h3>
+            <h3 id="value">Value: ${demoState.data.getValue()}</h3>
 
             <div id="buttons">
 
@@ -73,6 +73,13 @@ export class AsyncComponent2 extends LitStateElement {
 
             h3 {
                 font-size: 16px;
+            }
+
+            #status {
+                color: blue;
+            }
+
+            #value {
                 color: red;
             }